# perf(npm): deserialize cached npm packuments on blocking task