--- a/resolvers/npm_cache/lib.rs
+++ b/resolvers/npm_cache/lib.rs
@@ -308,7 +308,7 @@ impl<
       })
   }
 
-  pub fn load_package_info(
+  pub async fn load_package_info(
     &self,
     name: &str,
   ) -> Result<Option<SerializedCachedPackageInfo>, serde_json::Error> {
@@ -320,8 +320,9 @@ impl<
       Err(err) => return Err(serde_json::Error::io(err)),
     };
 
-    // todo(dsherret): deserialize in a blocking task
-    serde_json::from_slice(&file_bytes)
+    deno_unsync::spawn_blocking(move || serde_json::from_slice(&file_bytes))
+      .await
+      .unwrap()
   }
 
   pub fn save_package_info(