Noticed we missed doing this when looking through this code.

```sh
> cat package.json
{
  "dependencies": {
    "dax-sh": "^0.43.0",
    "next": "^15.3.1",
    "ts-morph": "^25.0.1"
  }
}
> hyperfine --warmup=1 "V:\deno\target\release\deno_main.exe install --no-lock" "V:\deno\target\release\deno.exe install --no-lock"                                
Benchmark 1: V:\deno\target\release\deno_main.exe install --no-lock
  Time (mean ± σ):     198.5 ms ±   3.2 ms    [User: 166.8 ms, System: 38.6 ms]                                                                                                                                                                                                                             
  Range (min … max):   192.7 ms … 203.3 ms    14 runs

Benchmark 2: V:\deno\target\release\deno.exe install --no-lock
  Time (mean ± σ):     146.4 ms ±   2.4 ms    [User: 178.7 ms, System: 133.0 ms]
  Range (min … max):   142.0 ms … 150.0 ms    19 runs

Summary
  V:\deno\target\release\deno.exe install --no-lock ran
    1.36 ± 0.03 times faster than V:\deno\target\release\deno_main.exe install --no-lock
```